% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category7-package.R
\docType{package}
\name{category7}
\alias{category7-package}
\alias{category7}
\title{category7: A Package for Categorical Data Analysis in S7}
\description{
A package designed for categorical data analysis built at top of S7, and subsequently an extension for `statim` package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/s7-stats/category7}
  \item Report bugs at \url{https://github.com/s7-stats/category7/issues}
}

}
\author{
\strong{Maintainer}: Joshua Marie \email{joshua.marie.k@gmail.com}

}
\keyword{internal}
