% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var-id.R
\name{cont_tab}
\alias{cont_tab}
\title{Contingency table of two variables}
\usage{
cont_tab(cat1, cat2, strict = TRUE)
}
\arguments{
\item{cat1}{The first categorical variable of the contingency table.}

\item{cat2}{The second categorical variable of the contingency table.}

\item{strict}{If \code{TRUE} (default), only a single variable is allowed for both
\code{cat1} and \code{cat2}. Set to \code{FALSE} to allow multiple variables, producing a
contingency table for each pair.}
}
\value{
An \code{cont_tab} / \code{var_id} S7 object.
}
\description{
\code{cont_tab()} creates a \code{cont_tab} model ID that reads as "the contingency table of
\code{cat1} and \code{cat2}".
}
\examples{
# Bare names
set.seed(123)
sex = sample(rep(c("Male", "Female"), each = 50))
species = sample(c("dog", "cat", "bird"), length(sex), replace = TRUE)
cont_tab(cat1, cat2)

}
